+++ demo/debug_script_01_uart_connection.py
@@ -22,8 +22,13 @@ print("---")
 # initiate the TMC_2209 class
 # use your pins for pin_en, pin_step, pin_dir here
 #-----------------------------------------------------------------------
-tmc = TMC_2209(21, 16, 20, skip_uart_init=True)
-
+if BOARD == "RASPBERRY_PI":
+    tmc = TMC_2209(21, 16, 20, skip_uart_init=True)
+elif BOARD == "NVIDIA_JETSON":
+    tmc = TMC_2209(13, 6, 5, serialport="/dev/ttyTHS1", skip_uart_init=True)
+else:
+    # just in case
+    tmc = TMC_2209(21, 16, 20, skip_uart_init=True)