+++ src/TMC_2209/_TMC_2209_test.py
@@ -9,10 +9,7 @@ TMC_2209 stepper driver communication module
 """
 
 import time
-try:
-    from RPi import GPIO
-except:
-    from Mock import GPIO
+from ._TMC_2209_GPIO_board import GPIO
 from ._TMC_2209_logger import Loglevel
 from ._TMC_2209_move import MovementAbsRel, MovementPhase
 from . import _TMC_2209_reg as tmc_reg
@@ -104,8 +101,19 @@ def test_uart(self):
     snd = result[0]
     rtn = result[1]
 
+    status = True
+
     self.tmc_logger.log(f"length snd: {len(snd)}", Loglevel.DEBUG)
     self.tmc_logger.log(f"length rtn: {len(rtn)}", Loglevel.DEBUG)
+
+
+    self.tmc_logger.log("complete messages:", Loglevel.DEBUG)
+    self.tmc_logger.log(str(snd.hex()), Loglevel.DEBUG)
+    self.tmc_logger.log(str(rtn.hex()), Loglevel.DEBUG)
+
+    self.tmc_logger.log("just the first 4 bits:", Loglevel.DEBUG)
+    self.tmc_logger.log(str(snd[0:4].hex()), Loglevel.DEBUG)
+    self.tmc_logger.log(str(rtn[0:4].hex()), Loglevel.DEBUG)
 
     if len(rtn)==12:
         self.tmc_logger.log("""the Raspberry Pi received the sended
@@ -113,12 +121,15 @@ def test_uart(self):
     elif len(rtn)==4:
         self.tmc_logger.log("the Raspberry Pi received only the sended bits",
                             Loglevel.INFO)
+        status = False
     elif len(rtn)==0:
         self.tmc_logger.log("the Raspberry Pi did not receive anything",
                             Loglevel.INFO)
+        status = False
     else:
         self.tmc_logger.log(f"the Raspberry Pi received an unexpected amount of bits: {len(rtn)}",
                             Loglevel.INFO)
+        status = False
 
     if snd[0:4] == rtn[0:4]:
         self.tmc_logger.log("""the Raspberry Pi received exactly the bits it has send.
@@ -126,16 +137,13 @@ def test_uart(self):
     else:
         self.tmc_logger.log("""the Raspberry Pi did not received the bits it has send.
                     the first 4 bits are different""", Loglevel.INFO)
+        status = False
 
-
-    self.tmc_logger.log("complete", Loglevel.DEBUG)
-    self.tmc_logger.log(str(snd.hex()), Loglevel.DEBUG)
-    self.tmc_logger.log(str(rtn.hex()), Loglevel.DEBUG)
-
-    self.tmc_logger.log("just the first 4 bits", Loglevel.DEBUG)
-    self.tmc_logger.log(str(snd[0:4].hex()), Loglevel.DEBUG)
-    self.tmc_logger.log(str(rtn[0:4].hex()), Loglevel.DEBUG)
-
+    self.tmc_logger.log("---")
+    if status:
+        self.tmc_logger.log("UART connection: OK")
+    else:
+        self.tmc_logger.log("UART connection: not OK")
 
     self.tmc_logger.log("---")
     return True